update to version 05.00.02, package rename, remove private repo from pom.xml
