+++ src/main/java/one/microstream/demo/bookstore/data/Customers.java
@@ -12,7 +12,7 @@ import java.util.stream.Stream;
 import one.microstream.demo.bookstore.BookStoreDemo;
 import one.microstream.demo.bookstore.util.concurrent.ReadWriteLocked;
 import one.microstream.persistence.types.Persister;
-import one.microstream.storage.types.EmbeddedStorageManager;
+import one.microstream.storage.embedded.types.EmbeddedStorageManager;
 
 /**
  * All registered customers of this company.
