+++ src/main/java/one/microstream/demo/bookstore/data/RandomDataGenerator.java
@@ -27,7 +27,7 @@ import org.rapidpm.dependencies.core.logger.HasLogger;
 import com.github.javafaker.Faker;
 
 import one.microstream.demo.bookstore.BookStoreDemo;
-import one.microstream.storage.types.EmbeddedStorageManager;
+import one.microstream.storage.embedded.types.EmbeddedStorageManager;
 
 /**
  * Random data generator for the {@link BookStoreDemo}'s {@link Data} root.
